requirements.txt
# Core ML/DL
torch>=2.1.0
pytorch-lightning>=2.1.0
torchmetrics>=1.2.0

# State-Space Models
mamba-ssm>=1.0.0
causal-conv1d>=1.1.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0  # for Parquet
polars>=0.19.0  # fast alternative to pandas

# Export & Deployment
onnx>=1.15.0
onnxruntime-gpu>=1.16.0
tensorrt>=8.6.0  # requires NVIDIA setup

# Serving
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0

# Visualization & Monitoring
matplotlib>=3.8.0
seaborn>=0.13.0
tensorboard>=2.15.0
wandb>=0.16.0  # optional cloud logging

# Config Management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
tqdm>=4.66.0
scikit-learn>=1.3.0
rich>=13.7.0  # beautiful CLI output
typer>=0.9.0  # CLI builder
